+++ config/dev.exs
@@ -55,3 +55,9 @@ config :td_dq, qc_types_file: "static/qc_types_dev.json"
 config :td_dq, qr_types_file: "static/qr_types_dev.json"
 
 config :td_perms, redis_host: "localhost"
+
+config :td_dq, :elasticsearch,
+  search_service: TdDq.Search,
+  es_host: "localhost",
+  es_port: 9200,
+  type_name: "doc"