+++ lib/td_dq/search/indexer.ex
@@ -1,0 +1,15 @@
+defmodule TdDq.Search.Indexer do
+  @moduledoc """
+    Manages elasticsearch indices
+  """
+  alias TdDq.ESClientApi
+  alias TdDq.Search
+  alias TdDq.Search.RuleMappings
+
+  def reindex(index_name, items) do
+    ESClientApi.delete!(index_name)
+    mapping = RuleMappings.get_mappings()
+    %{status_code: 200} = ESClientApi.put!(index_name, mapping |> Poison.encode!())
+    Search.put_bulk_search(items)
+  end
+end