+++ lib/td_dq/search/searchable.ex
@@ -1,0 +1,8 @@
+defmodule TdDq.Searchable do
+  @moduledoc """
+   Defines functions that must be implemented when having Searchable behaviour
+  """
+
+  @callback search_fields(any) :: any
+  @callback index_name() :: any
+end